requirements.txt
Django==1.5
MySQL-python==1.2.4
Pillow==2.2.1
South==0.8.4
django-bower==4.8.1
django-ckeditor==4.0.2
django-endless-pagination==2.0
django-haystack==2.1.0
django-taggit==0.11.1
django-taggit-templatetags==0.4.6dev
django-templatetag-sugar==0.1
django-updown==0.4.0
django-wysiwyg==0.5.1
flake8==2.1.0
mccabe==0.2.1
pep8==1.4.6
pyflakes==0.7.3
pytz==2013.8
requests==2.1.0
six==1.4.1
wsgiref==0.1.2
